Soft fix for duplicate work
The pow_evaluator failed to check if the same work had already been
applied in a given block. This could allow multiple miners to use the
same work until the difficulty invalidated the work. End result being
that the same work could be used up to 4 times at the nominal
difficutly, up to 8 times at twice the difficulty, and up to 12 times
at 16 times the difficulty.

This fix will cause anyone running the patch to reject duplicate work
while they are producing, but allow duplicate work included by other
miners.  This limits the ability of attackers exploiting the bug to
those who happen to produce a proof of work on block they are also
signing. This increases the difficulty of exploiting the loophole
20x.  This patch will also prevent duplicate work from propagating
across the P2P network.